First vue 3 app start..plus some stuff already.
Ok who'd have thought, yet here I am, learning 3rd of the big 3 frameworks. So how this shit works? So far...very different from Angular and React.

Vue app is generally composed of 3 things, no shocker - HTML, JS and CSS.
You import in the Vue script, create new Vue app object and then mount it on html element of your choosing. Anything inside that object will be acessible to Vue from then on.

let app = Vue.createApp({options object}); // create react app
let vm = app.mount('#vue_app'); //mount it

Ok here's where it starts getting freaky.
If you inspect Vue and app objects you're gonna see a bit of Proxy objects. And by bit I mean a fuckton. Get used to it, it's gonna stay that way. Usually you'll find the meat you need in their target objects.
Since you know what proxy is and how it works, ain't gonna explain it here.

Now on to options object passed to the Vue.createApp() call.
You'd expect passing an object with stuff inside and being done with it, right? WRONG!

Instead you gotta pass factory function and have it cough up the object in return statement. Why? How the fuck would I know? This is lesson one, not ending recap! Anyhoo here we passed data object with..yeah - data! Here it's msg property with welcome message.

You will find it in mounted app object (vm variable in this case) in two places. One is directly on it as it's property. That one is an alias, for ease of access.
The actual one is stored in $data. Both references point to same object so change it using any of it, other will reflect the change.

While we're on the money here ($), the $ is a Vue convention for naming it's properties. Any native Vue properties start with it.
One thing of note - the property "$" with nothing else is an object with some Vue internal stuff and should be left alone if you're not 100% confident about what you want from it and how to do it to not fuck it up.

Another small warning is the $el property. It points to child of the app's root element. But if the root element has more than one it starts going stupid (says "#text"). For now leave it. Better ways to do it will be introduced as we progress.

Whew! That's a shitton of text here. Nuff for one commit. More to come!