Setup function.
data function is not the only way to introduce new properties to the app's root component instance.
Alternative method for doing this is using a Vue 3 feature - setup function, which passes it's payload pretty much the same way as data - by returning object with new properies inside it.

The difference between the two lies in $data property. Setup properties are not added there, nor do they become proxies. They exist only as direct children of root component instance, and are not reactive by default. Changing their value will not cause html to be re-rendered to keep up with times.

Later that flaw will be tackled. For now it's important to realize that stuff displayed in moustache braces can have more than just one source.
And while on the subject - it's also good habit to keep stuff inside them as simple as possible. While lots of javascript expressions will work in them, it's best this pre-processing is done in other places, with moustache braces being used just to display result in cleanest way possible.