Directives and v-bind.
Directives serve the purpose of making html tag attributes dynamic and binding them to Vue logic and data.

A directive consists of 3 parts:

<input type="text" :value='msg' v-bind:disabled='true' />

First is the name/type of the directive - directives are a Vue thing, so you'll learn them as we move along.

Second the attribute it'll control. The moment this attribute is attached to a directive it stops being an HTML attribute and becomes a Vue attribute, which is to be interacted with Vue/javascript syntax.

Lastly is the value..but as mentioned above, we're not in Kansas anymore, but in Vue. The quotation marks holding the value are no longer string delimiters. Instead they are to hold javascript expression, and just as in case with moustache tags (the double curly braces), they have Vue's app root component scope, but without any access to window.

So you're free to snatch data passed with data object and place it here.

v-bind is first directive we'll learn here.
The : is a short-hand for v-bind btw if you wondered about that part:
<input (...) :value='msg' (...)

The v-bind let's you control the following attribute with vue and also  does one way data binding. If you change the value of js data it uses, the attribute in html will automatically update itself.