Conditional rendering directives
v-show takes in an expression that's either true or false.
if it's false, the tag with this directive will receive the style of
display: none. If set to true, it does nothing.

v-if is pretty much same as v-show with two major differences. If the passed expression evaluates to false, then the tag with this directive is not rendered at all, rather than just hidden with css.

Second difference is that it can be followed by v-else-if and v-else directives.

v-else-if - the name speaks for itself. You use that to do more if checks for specific conditions after the initial v-if.

v-else - once again pretty self explanatory. Always placed last (if used at all), it renders it's contents if prior if checks failed.

All of v-if type directives that belong to one logical block must be adjacent to eathother. You cannot do v-if, followed by tag with no if directives and then do another part of the if directive block. For example:

...
data: {display:false}

...
<p v-if='display'>You don't get to see me</p>
<p>Rando garbo</p>
<p v-else>I'm here</p>

This will error out. Second paragraph has no v-if-else breaking the conditional directive block.